Starting to change HttpObserver
The HttpObserver that detect connection to IceScrum is changing to
do this in three steps. The second step will be commited later. So,
let's start with the documentation. First, the code will check the
URL in the format used by the task of the tool partially. Not need
check all URL in the start and I drop the groups too. The second
step is verify if it is a sprintPlan URL. If it is a sprintPlan,
so it will drop the connection. This allow me to stay in the same
point of the page, but change the location. Finally, the third step
is check if the URL is a task URL. If it's true then the browser
will open the dialog and drop the connection. If the step one fail
or in step three is not a task URL, so the connection will be allowed.